Restore created_at sort in extraction helpers, remove duplicates from apiClient
