+++ src/services/apiClient.ts
@@ -1,12 +1,5 @@
 import { api } from "@/api/client";
-import type {
-  Reference,
-  SearchResult,
-  Enhancement,
-  EnhancementContent,
-  BibliographicMetadataEnhancement,
-  LinkedDataEnhancement,
-} from "@/types/models";
+import type { Reference, SearchResult } from "@/types/models";
 
 export interface SearchFilters {
   page?: number;
@@ -41,35 +34,4 @@ export async function getReference(id: string): Promise<Reference | null> {
     `/v1/references/?identifier=${encodeURIComponent(id)}`,
   );
   return results[0] ?? null;
-}
-
-function extractEnhancement<T extends EnhancementContent>(
-  reference: Reference,
-  enhancementType: T["enhancement_type"],
-): T | null {
-  if (!reference.enhancements) return null;
-  const matches = reference.enhancements.filter(
-    (e): e is Enhancement & { content: T } =>
-      e.content.enhancement_type === enhancementType,
-  );
-  if (matches.length === 0) return null;
-  const sorted = matches.sort((a, b) =>
-    (a.created_at ?? "").localeCompare(b.created_at ?? ""),
-  );
-  return sorted[sorted.length - 1].content;
-}
-
-export function extractBibliographic(
-  reference: Reference,
-): BibliographicMetadataEnhancement | null {
-  return extractEnhancement<BibliographicMetadataEnhancement>(
-    reference,
-    "bibliographic",
-  );
-}
-
-export function extractLinkedData(
-  reference: Reference,
-): LinkedDataEnhancement | null {
-  return extractEnhancement<LinkedDataEnhancement>(reference, "linked_data");
 }
