+++ src/services/referenceUtils.ts
@@ -1,33 +1,41 @@
 import type {
   Reference,
   Enhancement,
+  EnhancementContent,
   BibliographicMetadataEnhancement,
   LinkedDataEnhancement,
   ExternalIdentifier,
 } from "@/types/models";
 
+function extractEnhancement<T extends EnhancementContent>(
+  reference: Reference,
+  enhancementType: T["enhancement_type"],
+): T | null {
+  if (!reference.enhancements) return null;
+  const matches = reference.enhancements.filter(
+    (e): e is Enhancement & { content: T } =>
+      e.content.enhancement_type === enhancementType,
+  );
+  if (matches.length === 0) return null;
+  const sorted = matches.sort((a, b) =>
+    (a.created_at ?? "").localeCompare(b.created_at ?? ""),
+  );
+  return sorted[sorted.length - 1].content;
+}
+
 export function extractBibliographic(
   reference: Reference,
 ): BibliographicMetadataEnhancement | null {
-  if (!reference.enhancements) return null;
-  const matches = reference.enhancements.filter(
-    (e): e is Enhancement & { content: BibliographicMetadataEnhancement } =>
-      e.content.enhancement_type === "bibliographic",
+  return extractEnhancement<BibliographicMetadataEnhancement>(
+    reference,
+    "bibliographic",
   );
-  if (matches.length === 0) return null;
-  return matches[matches.length - 1].content;
 }
 
 export function extractLinkedData(
   reference: Reference,
 ): LinkedDataEnhancement | null {
-  if (!reference.enhancements) return null;
-  const matches = reference.enhancements.filter(
-    (e): e is Enhancement & { content: LinkedDataEnhancement } =>
-      e.content.enhancement_type === "linked_data",
-  );
-  if (matches.length === 0) return null;
-  return matches[matches.length - 1].content;
+  return extractEnhancement<LinkedDataEnhancement>(reference, "linked_data");
 }
 
 export function extractDoi(
