+++ tests/services/apiClient.test.ts
@@ -1,18 +1,7 @@
 import { vi } from "vitest";
 import { api } from "@/api/client";
-import {
-  searchReferences,
-  getReference,
-  extractBibliographic,
-  extractLinkedData,
-} from "@/services/apiClient";
-import type {
-  Reference,
-  SearchResult,
-  Enhancement,
-  BibliographicMetadataEnhancement,
-  LinkedDataEnhancement,
-} from "@/types/models";
+import { searchReferences, getReference } from "@/services/apiClient";
+import type { Reference, SearchResult } from "@/types/models";
 
 vi.mock("@/api/client", () => ({
   api: {
@@ -65,14 +54,6 @@ describe("searchReferences", () => {
     expect(params.getAll("annotation")).toEqual(["foo", "bar"]);
   });
 
-  test("appends multiple sort params", async () => {
-    mockedGet.mockResolvedValue(result);
-    await searchReferences("test", { sort: ["date", "-relevance"] });
-    const url = mockedGet.mock.calls[0][0];
-    const params = new URLSearchParams(url.split("?")[1]);
-    expect(params.getAll("sort")).toEqual(["date", "-relevance"]);
-  });
-
   test("returns the result from api.get", async () => {
     mockedGet.mockResolvedValue(result);
     const res = await searchReferences("test");
@@ -109,144 +90,5 @@ describe("getReference", () => {
     expect(mockedGet).toHaveBeenCalledWith(
       "/v1/references/?identifier=doi%3A10.1000%2Fabc123",
     );
-  });
-});
-
-function makeEnhancement(
-  content: Enhancement["content"],
-  createdAt: string | null = null,
-): Enhancement {
-  return {
-    id: null,
-    reference_id: "ref-1",
-    source: "test",
-    visibility: "public",
-    robot_version: null,
-    derived_from: null,
-    created_at: createdAt,
-    content,
-  };
-}
-
-function makeRef(enhancements: Enhancement[] | null): Reference {
-  return {
-    id: "ref-1",
-    visibility: "public",
-    identifiers: null,
-    enhancements,
-  };
-}
-
-describe("extractBibliographic", () => {
-  test("returns null when enhancements is null", () => {
-    expect(extractBibliographic(makeRef(null))).toBeNull();
-  });
-
-  test("returns null when no bibliographic enhancement exists", () => {
-    const ref = makeRef([
-      makeEnhancement({
-        enhancement_type: "abstract",
-        process: "x",
-        abstract: "y",
-      }),
-    ]);
-    expect(extractBibliographic(ref)).toBeNull();
-  });
-
-  test("returns bibliographic content when present", () => {
-    const bib: BibliographicMetadataEnhancement = {
-      enhancement_type: "bibliographic",
-      authorship: null,
-      cited_by_count: 5,
-      created_date: null,
-      updated_date: null,
-      publication_date: null,
-      publication_year: 2024,
-      publisher: null,
-      title: "Test",
-      pagination: null,
-      publication_venue: null,
-    };
-    const ref = makeRef([makeEnhancement(bib, "2026-01-01T00:00:00Z")]);
-    expect(extractBibliographic(ref)).toBe(bib);
-  });
-
-  test("returns the most recent bibliographic enhancement by created_at", () => {
-    const bib1: BibliographicMetadataEnhancement = {
-      enhancement_type: "bibliographic",
-      authorship: null,
-      cited_by_count: null,
-      created_date: null,
-      updated_date: null,
-      publication_date: null,
-      publication_year: 2020,
-      publisher: null,
-      title: "Older",
-      pagination: null,
-      publication_venue: null,
-    };
-    const bib2: BibliographicMetadataEnhancement = {
-      enhancement_type: "bibliographic",
-      authorship: null,
-      cited_by_count: null,
-      created_date: null,
-      updated_date: null,
-      publication_date: null,
-      publication_year: 2024,
-      publisher: null,
-      title: "Newer",
-      pagination: null,
-      publication_venue: null,
-    };
-    const ref = makeRef([
-      makeEnhancement(bib2, "2026-01-01T00:00:00Z"),
-      makeEnhancement(bib1, "2025-01-01T00:00:00Z"),
-    ]);
-    expect(extractBibliographic(ref)).toBe(bib2);
-  });
-});
-
-describe("extractLinkedData", () => {
-  test("returns null when enhancements is null", () => {
-    expect(extractLinkedData(makeRef(null))).toBeNull();
-  });
-
-  test("returns null when no linked_data enhancement exists", () => {
-    const ref = makeRef([
-      makeEnhancement({
-        enhancement_type: "abstract",
-        process: "x",
-        abstract: "y",
-      }),
-    ]);
-    expect(extractLinkedData(ref)).toBeNull();
-  });
-
-  test("returns linked_data content when present", () => {
-    const ld: LinkedDataEnhancement = {
-      enhancement_type: "linked_data",
-      vocabulary_uri: "http://example.com",
-      data: { key: "value" },
-    };
-    const ref = makeRef([makeEnhancement(ld, "2026-01-01T00:00:00Z")]);
-    expect(extractLinkedData(ref)).toBe(ld);
-  });
-
-  test("returns the most recent linked_data enhancement by created_at", () => {
-    const ld1: LinkedDataEnhancement = {
-      enhancement_type: "linked_data",
-      vocabulary_uri: "http://older.com",
-      data: { n: 1 },
-    };
-    const ld2: LinkedDataEnhancement = {
-      enhancement_type: "linked_data",
-      vocabulary_uri: "http://newer.com",
-      data: { n: 2 },
-    };
-    const ref = makeRef([
-      makeEnhancement(ld2, "2026-02-01T00:00:00Z"),
-      makeEnhancement(ld1, "2025-01-01T00:00:00Z"),
-    ]);
-    expect(extractLinkedData(ref)).toBe(ld2);
   });
 });
